+++ pyxel/models/data_processing/__init__.py
@@ -9,6 +9,7 @@
 
 # flake8: noqa
 from .statistics import statistics
+from .remove_cosmic_rays import remove_cosmic_rays
 from .source_extractor import extract_roi_to_xarray, plot_roi
 from .mean_variance import mean_variance
 from .linear_regression import linear_regression